refactor(theme,ui): sidebar token, terminal trim, paused-card icon, lint
- Add `--color-sidebar` token (Carbon #272930 — same as raised, so the
  left rail clearly lifts off content). Light variants map to their
  panel-gray. Sidebar.tsx + Settings sidebar use `bg-sidebar` so only
  the rail brightens in dark; modal/header/feed `bg-panel` unchanged.
- Drop the two light terminal palettes (Solarized Light + Catppuccin
  Latte). Terminal Theme picker is now Runner / Catppuccin Mocha /
  Solarized Dark — light terminals didn't pair well with TUIs that
  paint their own dark cards.
- Side-panel toggle: remove the dark border on the closed-state
  button, split icons into open + closed states (`PanelRight` for
  open, `PanelRightDashed` for closed) — both Obsidian-flavored, no
  chevrons. Applied in both RunnerChat and MissionWorkspace.
- Paused-card icon swapped `PowerOff` → `Pause`; cleaner semantic
  match to the "Chat paused" title.
- Lint: split `isFreshSpawn` + `chunkIndicatesTuiReady` out of
  SessionEndedOverlay.tsx into `lib/sessionLifecycle.ts` so the
  components file exports only components (Fast-Refresh-friendly).
  `useUpdate` stays co-located with its provider, with an inline
  rule disable explaining why.
